+++ src/config.ts
@@ -1,3 +1,16 @@
+export interface ConstructorData {
+    contractName: string;
+}
+
+export interface PositionsConstructorData extends ConstructorData {
+    tokenContract: string;
+    ratesContract: string;
+}
+
+export interface LiquidatorConstructorData extends ConstructorData {
+    positionsContract: string;
+}
+
 export const DECIMALS: { [key: string]: number } = {
     EOS: 4,
     PBTC: 8,